+++ src/utils/react-hooks.js
@@ -1,0 +1,6 @@
+import { useLocation } from "@reach/router"
+import { getNodeFilterByPath } from "./functions-non-es6"
+
+export function useNodeFilter() {
+  return getNodeFilterByPath(useLocation().pathname)
+}