+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Filtering by path prefix. E.g we can now filter by `/DE/`.
 - CHANGELOG
 
 ### Changed
 
+- Fix `gatsby-plugin-image` errors
 - `package.json` with `overrides` of `gatsby` -> `gatsby-script` to `1.4.0`
 
 ## [0.1.0] - 2023-13-31